+++ node/resolvers/profile/index.ts
@@ -1,8 +1,7 @@
-import { IOContext } from '@vtex/api'
 import { path } from 'ramda'
 import { MutationSaveAddressArgs } from 'vtex.store-graphql'
 
-import type { IdentityDataSource, User } from '../../dataSources/identity'
+import type { User } from '../../dataSources/identity'
 import fieldR from './fieldResolvers'
 import {
   createAddress,
@@ -18,28 +17,32 @@ const TRUE = 'True'
 const FALSE = 'False'
 
 interface CheckUserAuthorizationParams {
-  identity: IdentityDataSource
-  storeUserAuthToken: IOContext['storeUserAuthToken']
+  context: Context
   email: string
 }
 
 const checkUserAuthorization = async ({
-  identity,
-  storeUserAuthToken,
+  context,
   email,
 }: CheckUserAuthorizationParams): Promise<boolean> => {
-  let validUser = !!storeUserAuthToken
-  let userTokenData: Partial<User> | null = { user: '' }
+  const {
+    vtex: { storeUserAuthToken, account },
+    dataSources: { identity },
+  } = context
 
-  if (storeUserAuthToken) {
-    userTokenData = await identity.getUserWithToken(storeUserAuthToken)
-  }
+  let validUser = !!storeUserAuthToken
+
+  const tokenUser = await identity.getUserWithToken(storeUserAuthToken ?? '')
 
   validUser =
-    validUser && !!userTokenData && userTokenData?.user?.length === email.length
+    validUser &&
+    Boolean(tokenUser) &&
+    'id' in tokenUser &&
+    tokenUser.account === account &&
+    tokenUser.user.length === email.length
 
   for (let i = 0; i < email.length; i++) {
-    if (email[i] !== userTokenData?.user?.[i]) {
+    if (validUser && email[i] !== (tokenUser as User)?.user?.[i]) {
       validUser = false
     }
   }
@@ -95,7 +98,9 @@ export const mutations = {
       isNewsletterOptIn: optIn,
     }
 
-    const userProfile = await profile.getProfileInfo({ email, userId: '' })
+    const userProfile = await profile
+      .getProfileInfo({ email, userId: '' })
+      .catch(() => ({}))
 
     if (fields) {
       const { name, phone, bindingId, bindingUrl } = fields
@@ -129,11 +134,8 @@ export const mutations = {
     let updateData = true
 
     if (userProfile.createdIn) {
-      const { storeUserAuthToken } = context.vtex
-
       updateData = await checkUserAuthorization({
-        identity: context.dataSources.identity,
-        storeUserAuthToken,
+        context,
         email,
       })
     }