+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Use `storeUserAuthToken` to call new identity API.
+- Check if user is call center operator on withCurrentProfile.
 
 ## [2.149.1] - 2021-12-28
 ### Removed